+++ test/calcpace/test_pace_calculator.rb
@@ -12,11 +12,14 @@ class TestPaceCalculator < Minitest::Test
   def test_list_races
     races = @calc.list_races
     assert_kind_of Hash, races
-    assert_equal 4, races.size
+    assert_equal 7, races.size
     assert_equal 5.0, races['5k']
     assert_equal 10.0, races['10k']
     assert_equal 21.0975, races['half_marathon']
     assert_equal 42.195, races['marathon']
+    assert_equal 1.60934, races['1mile']
+    assert_equal 8.04672, races['5mile']
+    assert_equal 16.0934, races['10mile']
   end
 
   def test_race_time_with_numeric_pace